move supported noises
